1. TX management system was redeveloped. 2. EntityTests were updated.
