+++ entity-store/src/main/kotlin/jetbrains/exodus/entitystore/orientdb/OPersistentEntityStore.kt
@@ -39,10 +39,13 @@ class OPersistentEntityStore(
     private val dummyJobProcessor = object : MultiThreadDelegatingJobProcessor("dummy", 1) {}
     private val dummyStatistics = object : Statistics<Enum<*>>(arrayOf()) {}
     private val env = OEnvironment(databaseProvider.database, this)
-
+    private val currentTransaction = ThreadLocal<OStoreTransactionImpl>()
 
     override fun close() {
         //or it should be closed independently
+        currentTransaction.get()?.abort()
+
+        currentTransaction.remove()
         databaseProvider.close()
     }
 
@@ -53,10 +56,24 @@ class OPersistentEntityStore(
     }
 
     override fun beginTransaction(): StoreTransaction {
+        var currentTx = currentTransaction.get()
+
+        if (currentTx != null && currentTx.oTransaction.isActive) {
+            currentTx.oTransaction.begin()
+            return currentTx
+        }
+
         val session = databaseProvider.acquireSession()
-        session.activateOnCurrentThread()
         val txn = session.begin().transaction
-        return OStoreTransactionImpl(session, txn, this)
+
+        currentTx = OStoreTransactionImpl(session, txn, this)
+        currentTransaction.set(currentTx)
+
+        return currentTx
+    }
+
+    fun completeTx() {
+        currentTransaction.remove()
     }
 
     override fun beginExclusiveTransaction(): StoreTransaction {
@@ -67,12 +84,8 @@ class OPersistentEntityStore(
         return beginTransaction()
     }
 
-    override fun getCurrentTransaction(): StoreTransaction {
-        return OStoreTransactionImpl(
-            ODatabaseSession.getActiveSession(),
-            ODatabaseSession.getActiveSession().transaction,
-            this
-        )
+    override fun getCurrentTransaction(): StoreTransaction? {
+        return currentTransaction.get()
     }
 
     override fun getBackupStrategy(): BackupStrategy {
@@ -92,9 +105,10 @@ class OPersistentEntityStore(
         val txn = beginTransaction() as OStoreTransactionImpl
         try {
             executable.execute(txn)
-        } finally {
-            // if txn has not already been aborted in execute()
-            txn.activeSession.commit()
+            txn.commit()
+        } catch (e: Exception) {
+            txn.abort()
+            throw e
         }
     }
 
@@ -108,10 +122,12 @@ class OPersistentEntityStore(
         //i'm not sure about implementation
         val txn = beginTransaction() as OStoreTransactionImpl
         try {
-            return computable.compute(txn)
-        } finally {
-            // if txn has not already been aborted in execute()
-            txn.activeSession.commit()
+            val result = computable.compute(txn)
+            txn.commit()
+            return result
+        } catch (e: Exception) {
+            txn.abort()
+            throw e
         }
     }
 
@@ -156,7 +172,7 @@ class OPersistentEntityStore(
         val oldClass = computeInTransaction {
             val txn = it as OStoreTransaction
             txn.activeSession.metadata.schema.getClass(oldEntityTypeName)
-                ?: throw IllegalStateException("Class found by name $oldEntityTypeName")
+                ?: throw IllegalArgumentException("Class $oldEntityTypeName not found")
         }
         oldClass.setName(newEntityTypeName)
     }
@@ -171,7 +187,11 @@ class OPersistentEntityStore(
 
     override fun getStatistics() = dummyStatistics
 
-    override fun getAndCheckCurrentTransaction() = currentTransaction
+    override fun getAndCheckCurrentTransaction(): StoreTransaction {
+        val transaction = getCurrentTransaction()
+            ?: throw java.lang.IllegalStateException("EntityStore: current transaction is not set.")
+        return transaction
+    }
 
     override fun getCountsAsyncProcessor() = dummyJobProcessor
