+++ entity-store/src/test/java/jetbrains/exodus/entitystore/orientdb/OEntityStoreTestBase.java
@@ -1,0 +1,187 @@
+package jetbrains.exodus.entitystore.orientdb;
+
+import com.orientechnologies.orient.core.db.ODatabaseSession;
+import com.orientechnologies.orient.core.db.ODatabaseType;
+import com.orientechnologies.orient.core.db.OrientDB;
+import com.orientechnologies.orient.core.db.OrientDBConfig;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.Executors;
+
+import jetbrains.exodus.TestUtil;
+import jetbrains.exodus.entitystore.*;
+import jetbrains.exodus.util.IOUtil;
+
+import org.jetbrains.annotations.NotNull;
+import org.junit.Assert;
+
+public abstract class OEntityStoreTestBase extends TestBase {
+
+    private static final String TEMP_FOLDER = TestUtil.createTempDir().getAbsolutePath();
+
+    private OrientDB orientDB;
+    private PersistentEntityStore store;
+
+    protected boolean isPartiallyTornDown;
+    protected boolean shouldCleanopOnTearDown;
+    private String databaseFolder;
+
+
+    @Override
+    protected void setUp() throws Exception {
+        isPartiallyTornDown = false;
+        shouldCleanopOnTearDown = true;
+
+        orientDB = OrientDB.embedded(getClass().getSimpleName(), OrientDBConfig.defaultConfig());
+        if (orientDB.exists(getName())) {
+            orientDB.drop(getName());
+        }
+
+        orientDB.create(getName(), ODatabaseType.MEMORY, "admin", "admin", "admin");
+
+        openStore();
+    }
+
+    protected PersistentEntityStore openStore() {
+        return store = createStoreInternal(getDatabaseFolder());
+    }
+
+    protected String getDatabaseFolder() {
+        if (databaseFolder == null) {
+            databaseFolder = initTempFolder();
+        }
+        return databaseFolder;
+    }
+
+    protected PersistentEntityStore createStoreInternal(String dbTempFolder) {
+        return createStore(dbTempFolder);
+    }
+
+    public PersistentEntityStore createStore(String dbTempFolder) {
+        return new OPersistentEntityStore(new ODatabaseProvider() {
+            @NotNull
+            @Override
+            public String getDatabaseLocation() {
+                return "";
+            }
+
+            @NotNull
+            @Override
+            public OrientDB getDatabase() {
+                return orientDB;
+            }
+
+            @NotNull
+            @Override
+            public ODatabaseSession acquireSession() {
+                return orientDB.cachedPool(getName(), "admin", "admin").acquire();
+            }
+
+            @Override
+            public void close() {
+                orientDB.close();
+            }
+        }, getName(), Executors.newSingleThreadExecutor());
+    }
+
+    @Override
+    protected void tearDown() {
+        if (!isPartiallyTornDown) {
+            store.close();
+        }
+        if (shouldCleanopOnTearDown) {
+            cleanUp(store.getLocation());
+            databaseFolder = null;
+        }
+    }
+
+    public static String initTempFolder() {
+        // Configure temp folder for database
+        final String location = randomTempFolder();
+        final File tempFolder = new File(location);
+        if (!tempFolder.mkdirs()) {
+            Assert.fail("Can't create directory at " + location);
+        }
+        if (logger.isInfoEnabled()) {
+            logger.info("Temporary data folder created: " + location);
+        }
+        return tempFolder.getAbsolutePath();
+    }
+
+    public static void cleanUp(String location) {
+        final File tempFolder = new File(location);
+        if (logger.isInfoEnabled()) {
+            logger.info("Cleaning data folder: " + location);
+        }
+        IOUtil.deleteRecursively(tempFolder);
+        IOUtil.deleteFile(tempFolder);
+    }
+
+    public void transactional(
+            @NotNull final StoreTransactionalExecutable executable) {
+        store.executeInTransaction(executable);
+    }
+
+    public void transactionalExclusive(
+            @NotNull final StoreTransactionalExecutable executable) {
+        store.executeInExclusiveTransaction(executable);
+    }
+
+    public void transactionalReadonly(
+            @NotNull final StoreTransactionalExecutable executable) {
+        store.executeInReadonlyTransaction(executable);
+    }
+
+    protected final PersistentEntityStore getEntityStore() {
+        return store;
+    }
+
+    @Override
+    protected String getArtifactsPath() {
+        return "." + File.separatorChar + "testartifacts" + File.separatorChar;
+    }
+
+    @Override
+    protected File[] getFilesToBackup() {
+        final File tempFolder = new File(store.getLocation());
+        store.close();
+        isPartiallyTornDown = true;
+        return new File[]{tempFolder};
+    }
+
+    public ODatabaseSession acquireSession() {
+        return orientDB.cachedPool(getName(), "admin", "admin").acquire();
+    }
+
+    public static String randomTempFolder() {
+        return TEMP_FOLDER + Math.random();
+    }
+
+    protected void reinit() throws Exception {
+        shouldCleanopOnTearDown = false;
+        tearDown();
+        setUp();
+    }
+
+    @NotNull
+    private static StoreTransactionalExecutable wrap(
+            @NotNull final EntityStoreTestBase.PersistentStoreTransactionalExecutable executable) {
+        return txn -> executable.execute(
+                (PersistentStoreTransaction) txn);
+    }
+
+    public interface PersistentStoreTransactionalExecutable {
+
+        void execute(@NotNull final PersistentStoreTransaction txn);
+    }
+
+    public static List<Entity> toList(Iterable<Entity> it) {
+        final List<Entity> result = new ArrayList<>();
+        for (Entity entity : it) {
+            result.add(entity);
+        }
+        return result;
+    }
+}
